Fix (?) imports so that they are compile
Imports were relative to project root, but my compiler
seems to want the imports to be relative to the source
file the import is found in. Not sure if there is just
several different ways to compile a Go program.

For now, it seems I need to relativize these imports.